ai_reviews/requirements.txt
azure-ai-ml>=1.11.0
azure-identity>=1.15.0
